Added destroy action
Now is posible to delete a project(in the future a lot of things will
depend of the project (issues, test cases, etc) when this happend if the
project dies all it dependencies dies with it). Also added a link to the
project index view, when clicked the user will be taken to the create
new project form.